backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.0
reportlab==4.0.7
bleach==6.1.0
email-validator==2.1.0
gunicorn==21.2.0
requests==2.32.3
